backend/requirements.txt
flask
flask_sqlalchemy
alembic
python-dotenv
flask_wtf
wtforms
flask_jwt_extended
bcrypt